Executed refactoring in AprgProcessFiles module.

+++ AprgProcessFiles/AprgProcessFiles/src/RenameFiles/AlbaRenameFiles.cpp
@@ -20,15 +20,13 @@ void AlbaRenameFiles::renameWithIncreasingNumbers(
 {
     NumberToStringConverter converter;
     converter.setFillCharacter('0');
-    converter.setFieldWidth(fieldWidth);
+    converter.setFieldWidth(static_cast<int>(fieldWidth));
 
     ListOfPaths listOfFiles;
-    ListOfPaths listOfDirectories;
-    m_directoryPathHandler.findFilesAndDirectoriesOneDepth("*.*", listOfFiles, listOfDirectories);
+    ListOfPaths listOfDirectories;    m_directoryPathHandler.findFilesAndDirectoriesOneDepth("*.*", listOfFiles, listOfDirectories);
 
     unsigned int count=startingNumber;
-    for(string const& file : listOfFiles)
-    {
+    for(string const& file : listOfFiles)    {
         AlbaLocalPathHandler filePathHandler(file);
         filePathHandler.renameFile(prefix+converter.convert(count++)+string(".")+filePathHandler.getExtension());
     }
